Updated project notebook with out of distribution testing
We now have two GTZAN style datasets. The original on which we train and one made up of copyrighted music that cannot have been part of the original. This new dataset is for use in testing the generalization capabilities of the model to music in a different distribution than the training set.